Add support for use of ext4 in containers
By default, LXD won't allow an unpriviledged container
to format and mount a block device, so although a block
device can be passed to a container, its not possible
to use it from inside the container.

Add a configuration flag to turn on support for use of
ext4 within user namespaces, allow unpriviledged LXD
containers to format and mount ext4 formatted block
devices from inside of the container.

Change-Id: I8eace1e7b7d4db14a3dc9f82080ac1b0ca29e892