Increase subuid/subgid range for root user
To support use of distinct subuid/subgid ranges per LXD container,
the default range for the root user must be increased to support
> 1 running container in this configuration.

Increase subuid/subgid range to support 5000 containers with distinct
ranges.  Restart LXD daemon if idmap configuration changes, to ensure
that the full range of subid's are used.

Change-Id: I8b87dad736abaffdbd7afac090429790d3b03c96
Closes-Bug: 1648056