Implements support for snap installing LXD
This patchset adds support for performing a snap installation of LXD on
the nova-compute host that the LXD subordinate is added to.
There a corresponding changes in the nova-compute charm and in the
nova-lxd plugin for nova.  The reason is because some directories change
depending on whether LXD is snap-installed or not, and in the case of
nova-lxd, the location of the console.log.

Change-Id: I2b365891679463cafd3a325ce0caab9a20e5ab2e